+++ libs/mixer/src/tank_steer_strategy.cpp
@@ -6,7 +6,7 @@
 
 using namespace MIXER;
 
-MotorSpeeds TankSteerStrategy::mix(float velocity, float angularVelocity) {
+AckermannOutput TankSteerStrategy::mix(float velocity, float angularVelocity) {
     float left = velocity - angularVelocity;
     float right = velocity + angularVelocity;
     float abs_left = left >= 0 ? left : -left;
@@ -20,5 +20,6 @@ MotorSpeeds TankSteerStrategy::mix(float velocity, float angularVelocity) {
     const float speed_factor = 1.0f;
     float scaled_left = (left * speed_factor);
     float scaled_right = (right * speed_factor) * -1;
-    return {scaled_left, scaled_right, scaled_left, scaled_right};
+    
+    return {scaled_left, scaled_right, scaled_left, scaled_right, 0, 0};
 }
