+++ libs/statemanager/include/statemanager.h
@@ -41,7 +41,7 @@ namespace STATEMANAGER {
         // max speed factor - scale the speed of the motors down to this value
         static constexpr float SPEED_EXTENT = 1.0f;
 
-        void setSpeeds(MIXER::MotorSpeeds motorSpeeds) const;
+        void setSpeeds(MIXER::AckermannOutput motorSpeeds) const;
     };
 
 } // StateManager
